chore: consolidate tx warning component files
