+++ components/brave_wallet_ui/components/extension/confirm-transaction-panel/common/tx_warnings.styles.tsx
@@ -5,9 +5,37 @@
 
 import styled from 'styled-components'
 import * as leo from '@brave/leo/tokens/css'
+
+import Icon from '@brave/leo/react/icon'
 import Collapse from '@brave/leo/react/collapse'
 
-import { styledScrollbarMixin } from '../../../shared/style'
+// shared styles
+import { Row, styledScrollbarMixin } from '../../../shared/style'
+
+interface WarningBannerProps {
+  isCritical?: boolean
+}
+
+export const AlertIcon = styled(Icon).attrs<WarningBannerProps>((props) => ({
+  name: props.isCritical ? 'warning-circle-filled' : 'warning-triangle-filled'
+}))<WarningBannerProps>`
+  --leo-icon-size: 20px;
+  padding: 0px 16px;
+  color: ${(p) =>
+    p.isCritical
+      ? leo.color.systemfeedback.errorIcon
+      : leo.color.systemfeedback.warningIcon};
+`
+
+// TODO: match new collapse style
+export const WarningAlertRow = styled(Row)<WarningBannerProps>`
+  min-height: 44px;
+  color: ${leo.color.text.primary};
+  background-color: ${(p) =>
+    p.isCritical
+      ? leo.color.systemfeedback.errorBackground
+      : leo.color.systemfeedback.warningBackground};
+`
 
 export const WarningCollapse = styled(Collapse)<{
   isCritical?: boolean
